+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Reorganise ABI compilation process, source dependencies and `eth_defi.abi` folder layout\
 - In-house contracts are now compiled using [Foundry](https://book.getfoundry.sh/)
+- Add `VaultSpecificGenericAdapter.sol` for Enzyme
 - Add `eth_defi.enzyme.vault_controlled_vallet`